fix: moved formatDateTime to utils for reuse
